Renamed lots of code
