+++ lib/transito/marshaler/jruby/json.rb
