+++ lib/transito/rolling_cache.rb
