+++ spec/spec_helper.rb
@@ -12,19 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-unless File.exist?('../transit-format/examples/0.8/simple')
+unless File.exist?('../transito-format/examples/0.8/simple')
   puts <<-MSG
 Before you can run the rspec examples, you need to install the
-the https://github.com/cognitect/transit-format repo in a sibling
+the https://github.com/cognitect/transito-format repo in a sibling
 directory, e.g.
 
     cd ..
-    git clone https://github.com/cognitect/transit-format
+    git clone https://github.com/cognitect/transito-format
 
-That repo contains exemplars used by transit-ruby's rspec examples
-(in ../transit-format/examples/0.8/simple), so then you can:
+That repo contains exemplars used by transito-ruby's rspec examples
+(in ../transito-format/examples/0.8/simple), so then you can:
 
-    cd transit-ruby
+    cd transito-ruby
     rspec
 
 MSG
@@ -35,7 +35,7 @@ require 'json'
 require 'rspec'
 require 'date' # because wrong expects it but doesn't load it :/
 require 'wrong/adapters/rspec'
-require 'transit'
+require 'transito'
 require 'spec_helper-local' if File.exist?(File.expand_path('../spec_helper-local.rb', __FILE__))
 
 RSpec.configure do |c|
