+++ spec/transito/marshaler_spec.rb
