+++ spec/transito/reader_spec.rb
