+++ spec/transito/rolling_cache_spec.rb
